+++ py/custom_noise/__init__.py
@@ -1,8 +1,9 @@
-from . import noise_perlin
 from . import nodes
 
 NODE_CLASS_MAPPINGS = {
-    "OCSNoise PerlinSimple": noise_perlin.PerlinSimpleNode,
-    "OCSNoise PerlinAdvanced": noise_perlin.PerlinAdvancedNode,
+    "OCSNoise PerlinSimple": nodes.PerlinSimpleNode,
+    "OCSNoise PerlinAdvanced": nodes.PerlinAdvancedNode,
     "OCSNoise to SONAR_CUSTOM_NOISE": nodes.ToSonarNode,
+    "OCSNoise Conditioning": nodes.NoiseConditioningNode,
+    "OCSNoise OverrideSamplerNoise": nodes.SamplerNodeConfigOverride,
 }
