+++ py/custom_noise/noise_perlin.py
@@ -1,11 +1,12 @@
-import torch
-import math
 import itertools
+import math
 
-from .base import CustomNoiseItemBase, CustomNoiseNodeBase, NormalizeNoiseNodeMixin
+import torch
+
+from ..filtering import BLENDING_MODES
 from ..latent import normalize_to_scale
 from ..noise import scale_noise
-from ..filtering import BLENDING_MODES
+from .base import CustomNoiseItemBase, NormalizeNoiseNodeMixin
 
 # Perlin generation routines based on https://github.com/Extraltodeus/noise_latent_perlinpinpin which was based on https://gist.github.com/vadimkantorov/ac1b097753f217c5c11bc2ff396e0a57 which was based on https://github.com/pvigier/perlin-numpy
 
@@ -484,263 +485,3 @@ class PerlinItem(CustomNoiseItemBase):
             return scale_noise(noise, self.factor, normalized=normalized)
 
         return noise_sampler
-
-
-class PerlinAdvancedNode(CustomNoiseNodeBase, NormalizeNoiseNodeMixin):
-    DESCRIPTION = "Advanced Perlin noise generator, allows generating 2D or 3D Perlin noise. See the OCSNoise PerlinSimple node for less tuneable parameters."
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        result = super().INPUT_TYPES()
-        result["required"] |= {
-            "depth": (
-                "INT",
-                {
-                    "default": DEFAULTS.depth,
-                    "tooltip": "When non-zero, 3D perlin noise will be generated.",
-                },
-            ),
-            "detail_level": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.detail_level,
-                    "tooltip": "Controls the detail level of the noise when break_pattern is non-zero. No effect when using 100% raw Perlin noise.",
-                },
-            ),
-            "octaves": (
-                "INT",
-                {
-                    "default": DEFAULTS.octaves,
-                    "tooltip": "Generally controls the detail level of the noise. Each octave involves generating a layer of noise so there is a performance cost to increasing octaves.",
-                },
-            ),
-            "persistence": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("persistence"),
-                    "tooltip": "Controls how rough the generated noise is. Lower values will result in smoother noise, higher values will look more like Gaussian noise. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "lacunarity_height": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("lacunarity", 0),
-                    "tooltip": "Lacunarity controls the frequency multiplier between successive octaves. Only has an effect when octaves is greater than one. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "lacunarity_width": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("lacunarity", 1),
-                    "tooltip": "Lacunarity controls the frequency multiplier between successive octaves. Only has an effect when octaves is greater than one. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "lacunarity_depth": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("lacunarity", 2),
-                    "tooltip": "Lacunarity controls the frequency multiplier between successive octaves. Only has an effect when depth is non-zero and octaves is greater than one. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "res_height": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("res", 0),
-                    "tooltip": "Number of periods of noise to generate along an axis. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "res_width": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("res", 1),
-                    "tooltip": "Number of periods of noise to generate along an axis. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "res_depth": (
-                "STRING",
-                {
-                    "default": DEFAULTS.get_commasep("res", 2),
-                    "tooltip": "Number of periods of noise to generate along an axis. Only has an effect when depth is non-zero. Comma-separated list, multiple items will apply to octaves in sequence.",
-                },
-            ),
-            "break_pattern": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.break_pattern,
-                    "tooltip": "Applies a function to break the Perlin pattern, making it more like normal noise. The value is the blend strength, where 1.0 indicates 100% pattern broken noise and 0.5 indicates 50% raw noise and 50% pattern broken noise. Generally should be at least 0.9 unless you want to generate colorful blobs.",
-                },
-            ),
-            "initial_depth": (
-                "INT",
-                {
-                    "default": DEFAULTS.initial_depth,
-                    "tooltip": "First zero-based depth index the noise generator will return. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "wrap_depth": (
-                "INT",
-                {
-                    "default": DEFAULTS.wrap_depth,
-                    "tooltip": "If non-zero, instead of generating a new chunk of noise when the last slice is used will instead jump back to the specified zero-based depth index. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "max_depth": (
-                "INT",
-                {
-                    "default": DEFAULTS.max_depth,
-                    "tooltip": "Basically crops the depth dimension to the specified value (inclusive). Negative values start from the end, the default of -1 does no cropping. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "tileable_height": (
-                "BOOLEAN",
-                {
-                    "default": DEFAULTS.tileable[0],
-                    "tooltip": "Makes the specified dimension tileable.",
-                },
-            ),
-            "tileable_width": (
-                "BOOLEAN",
-                {
-                    "default": DEFAULTS.tileable[1],
-                    "tooltip": "Makes the specified dimension tileable.",
-                },
-            ),
-            "tileable_depth": (
-                "BOOLEAN",
-                {
-                    "default": DEFAULTS.tileable[2],
-                    "tooltip": "Makes the specified dimension tileable. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "blend": (
-                tuple(BLENDING_MODES.keys()),
-                {
-                    "default": "lerp",
-                    "tooltip": "Blending function used when generating Perlin noise. When set to values other than LERP may not work at all or may not actually generate Perlin noise.",
-                },
-            ),
-            "pattern_break_blend": (
-                tuple(BLENDING_MODES.keys()),
-                {
-                    "default": "lerp",
-                    "tooltip": "Blending function used to blend pattern broken noise with raw noise.",
-                },
-            ),
-            "depth_over_channels": (
-                "BOOLEAN",
-                {
-                    "default": DEFAULTS.depth_over_channels,
-                    "tooltip": "When disabled, each channel will have its own separate 3D noise pattern. When enabled, depth is multiplied by the number of channels and each channel is a slice of depth. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "pad_height": (
-                "INT",
-                {
-                    "default": DEFAULTS.pad[0],
-                    "min": 0,
-                    "tooltip": "Pads the specified dimension by the size. Equal padding will be added on both sides and cropped out after generation.",
-                },
-            ),
-            "pad_width": (
-                "INT",
-                {
-                    "default": DEFAULTS.pad[1],
-                    "min": 0,
-                    "tooltip": "Pads the specified dimension by the size. Equal padding will be added on both sides and cropped out after generation.",
-                },
-            ),
-            "pad_depth": (
-                "INT",
-                {
-                    "default": DEFAULTS.pad[2],
-                    "min": 0,
-                    "tooltip": "Pads the specified dimension by the size. Equal padding will be added on both sides and cropped out after generation. Only has an effect when depth is non-zero.",
-                },
-            ),
-            "initial_amplitude": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.initial_amplitude,
-                    "tooltip": "Controls the amplitude for the first octave.",
-                },
-            ),
-            "initial_frequency_height": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.initial_frequency[0],
-                    "tooltip": "Controls the frequency for the first octave for the this axis.",
-                },
-            ),
-            "initial_frequency_width": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.initial_frequency[1],
-                    "tooltip": "Controls the frequency for the first octave for the this axis.",
-                },
-            ),
-            "initial_frequency_depth": (
-                "FLOAT",
-                {
-                    "default": DEFAULTS.initial_frequency[2],
-                    "tooltip": "Controls the frequency for the first octave for the this axis.",
-                },
-            ),
-            "normalize": (
-                ("default", "forced", "off"),
-                {
-                    "tooltip": "Controls whether the output noise is normalized after generation.",
-                },
-            ),
-            "device": (
-                ("default", "cpu", "gpu"),
-                {
-                    "default": "default",
-                    "tooltip": "Controls what device is used to generate the noise. GPU noise may be slightly faster but you will get different results on different GPUs.",
-                },
-            ),
-        }
-        return result
-
-    @classmethod
-    def get_item_class(cls):
-        return PerlinItem
-
-
-class PerlinSimpleNode(PerlinAdvancedNode):
-    DESCRIPTION = "Simplified Perlin noise generator, allows generating 2D or 3D Perlin noise. See the OCSNoise PerlinAdvanced node for more tuneable parameters."
-
-    _COPY_KEYS = {
-        "factor",
-        "rescale",
-        "depth",
-        "detail_level",
-        "octaves",
-        "persistence",
-        "break_pattern",
-    }
-
-    @classmethod
-    def INPUT_TYPES(cls):
-        result = super().INPUT_TYPES()
-        orig_reqs = result["required"]
-        reqs = {k: v for k, v in orig_reqs.items() if k in cls._COPY_KEYS}
-        reqs["lacunarity"] = orig_reqs["lacunarity_height"]
-        reqs["res"] = orig_reqs["res_height"]
-        result["required"] = reqs
-        return result
-
-    @classmethod
-    def get_item_class(cls):
-        def wrapper(factor, *, lacunarity, res, **kwargs):
-            return PerlinItem(
-                factor,
-                lacunarity_height=lacunarity,
-                lacunarity_width=lacunarity,
-                lacunarity_depth=lacunarity,
-                res_height=res,
-                res_width=res,
-                res_depth=res,
-                **kwargs,
-            )
-
-        return wrapper
